Updated Report in support of ACATS-grading.
Made function Legal_File_Name 'separate'. This is to allow each test
to run in its own directory in Grading, which will mean that external
file names that need to be preserved between tests have to be one
level up.

This version has the same functionality as the original; Grading will
replace it to adjust the generated file names as required.

  * support/report.a: as above.